fix: DataLoader num_workers=0 to prevent CUDA fork deadlock; GLUE bfloat16 cast
- peat.py: set num_workers=0 in both DataLoader calls (SHA and final-train).
  num_workers>0 causes fork() after CUDA is initialised (from the prior model),
  which deadlocks PyTorch worker processes on Linux. The smoke-test was
  consistently hanging at modernbert-base SHA Round-0 epoch 1 (0%) because
  the prefetch thread had already loaded NomicBERT tensors into shared memory
  before the workers forked.

- eval.py: add .float() before .numpy() in extract_features so sklearn's
  LogisticRegression receives float32 arrays instead of bfloat16, fixing
  'Got unsupported ScalarType BFloat16' warnings and the resulting
  GLUE average: nan for all encoder models.

+++ Code/peat/eval.py
@@ -629,7 +629,7 @@ def evaluate_glue(model, tokenizer, model_tag: str,
                                 cls_hidden = outputs.last_hidden_state[:, 0, :]
                             else:
                                 cls_hidden = outputs.logits[:, 0, :]
-                    features_list.append(cls_hidden.cpu().numpy().flatten())
+                    features_list.append(cls_hidden.cpu().float().numpy().flatten())
                     labels.append(ex["label"])
                 return np.array(features_list), np.array(labels)
